+++ scilab2py/__init__.py
@@ -18,7 +18,7 @@ MAT files.  Usage is as simple as:
 
 
 __title__ = 'scilab2py'
-__version__ = '0.2'
+__version__ = '0.3'
 __author__ = 'Steven Silvester'
 __license__ = 'MIT'
 __copyright__ = 'Copyright 2014 Steven Silvester'