+++ HISTORY.rst
@@ -8,7 +8,8 @@ Release History
 - Allow keyword arguments in functions: `octave.plot([1,2,3], linewidth=2))`
   These are translated to ("prop", value) arguments to the function.
 - Add option to show plotting gui with `-g` flag in ScilabMagic.
-
+- Add ability to specify the Scilab executable as a keyword argument to
+  the Scilab2Py object.
 
 0.2.0 (2014-08-14)
 ++++++++++++++++++