+++ HISTORY.rst
@@ -3,7 +3,7 @@
 Release History
 ---------------
 
-0.3.0 (2014-xx-xx)
+0.3.0 (2014-08-23)
 ++++++++++++++++++
 - Allow keyword arguments in functions: `octave.plot([1,2,3], linewidth=2))`
   These are translated to ("prop", value) arguments to the function.